ceph-volume: lvm list's pretty report simplification
Users create osds with the ceph volume's lvm toolkit by giving
devices in the form of e.g. "/dev/sdX". When they ask ceph-volume
to list the available devices, they get nothing related to the disks
they used in the arguments (especially the block device). Even when
they try to use `ceph-volume lvm list /dev/sdX` they get an empty
report.

The code that creates the disks creates new volume groups for each
osd, so it is safe to assume that each physical volume corresponds
to a volume group and to a logical volume.

The output repetition is removed and the listing is now closer to
ceph-disk's listing output. The json format output is almost unchanged.

Signed-off-by: Theofilos Mouratidis <t.mour@cern.ch>